requirements.txt
alembic==1.3.2
aniso8601==8.0.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.1.0
numpy==1.18.1
pandas==0.25.3
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.4
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
WTForms==2.2.1
zipp==1.0.0
